Update MyRobot.cpp
Updated convert_bearing_to_degrees to an in_vector sequence that will work with compute_odometry. Also added green_identifier, blue_identifier, and green_drive for searching the endzone. These three functions have global variables that need to be declared in the .h and initialized before the while loop as zero. These variables have not been added in this update.